requirements/base.txt
azure-storage==0.36.0
coreapi==2.3.3
Django==2.2.2
django-allauth==0.41.0
django-cors-headers==3.2.1
django-extensions==2.2.8
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-framework-docs==0.1.7
djangorestframework==3.11.0
drf-nested-routers==0.91
psycopg2-binary==2.8.4

